requirements.txt
djongo
django
djangorestframework
django-allauth
psycopg2-binary
python-decouple
pylint-django
django-cors-headers
djangorestframework-simplejwt
gunicorn
whitenoise
psycopg2
dj-database-url
django
djangorestframework
requests
django-recurrence
Pillow